+++ src/App.test.jsx
@@ -2,8 +2,9 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import App from './App';
 
-test('renders title', () => {
-  const { getByText } = render(<App />);
-  const linkElement = getByText(/Pools/);
-  expect(linkElement).toBeInTheDocument();
+test('renders header & footer text', () => {
+  const { getAllByText } = render(<App />);
+  const elements = getAllByText(/Pools/);
+  expect(elements).toHaveLength(2);
+  expect(elements[0]).toBeInTheDocument();
 });