+++ src/components/ContentWrapper.jsx
@@ -1,9 +1,11 @@
 import React from 'react';
 import Content from './Content';
+import Footer from './Footer';
 
 const ContentWrapper = () => (
   <div id="content-wrapper" className="d-flex flex-column">
     <Content />
+    <Footer />
   </div>
 );