+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 
 ## [Unreleased]
 
-  - Introduce React Router, refs #4
   - Setup CI, refs #2
+  - Introduce React Router, refs #4
+  - Add website version to footer, refs #6
   - Display pool token price, refs #7